negociate

+++ C Piscine C Rush 02/ex00/include/ft_ab_types.h
@@ -6,7 +6,7 @@
 /*   By: jseo <jseo@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/31 15:42:59 by jseo              #+#    #+#             */
-/*   Updated: 2020/10/31 18:43:54 by jseo             ###   ########.fr       */
+/*   Updated: 2020/11/01 11:29:59 by jseo             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define FT_AB_TYPES_H
 
 typedef unsigned int	t_int;
-typedef unsigned int	t_long;
+typedef unsigned long	t_long;
 typedef unsigned char	t_char;
 
 #endif
